Extract OSGi setup into separate build script
